feat: i18n audit, split import/mcp pages into components
- Extract hardcoded strings from skills/new-version, skills/evolve,
  backup, and chat pages with full 5-language i18n support
- Split import/page.tsx (1050 lines) into 5 focused components:
  file-import-tab, github-import-tab, folder-import-tab,
  conflict-dialog, import-result-display
- Split mcp/page.tsx (818 lines) into 6 focused components:
  mcp-server-card, add-server-dialog, secret-config-dialog,
  secrets-section, tools-list, mcp-tool-card
- Add aria-label to published chat textarea
