+++ web/src/i18n/locales/en-US/chat.json
@@ -33,6 +33,7 @@
   "agentNotAvailable": "Agent Not Available",
   "builtinTools": "Built-in Tools",
   "selected": "selected",
+  "toolRequired": "(required)",
   "configuration": {
     "title": "Configuration",
     "agent": "Agent",
